refactor(music/entity): rename package api to model
