+++ core/music-reactive/src/main/java/com/github/durex/music/service/MusicService.java
@@ -6,7 +6,7 @@ import static com.github.durex.shared.exceptions.model.ErrorCode.ENTITY_NOT_FOUN
 import static com.github.durex.shared.exceptions.model.ErrorCode.SAVE_ERROR;
 import static com.github.durex.shared.exceptions.model.ErrorCode.UPDATE_ERROR;
 
-import com.github.durex.music.api.Music;
+import com.github.durex.music.model.Music;
 import com.github.durex.music.repository.MusicRepository;
 import com.github.durex.shared.exceptions.ApiException;
 import com.github.durex.sqlbuilder.enums.WildCardType;
