+++ core/music-reactive/src/main/java/com/github/durex/music/service/PlaylistService.java
@@ -5,9 +5,9 @@ import static com.github.durex.music.support.EntityConstants.MUSIC_NOT_UPDATED;
 import static com.github.durex.shared.exceptions.model.ErrorCode.ENTITY_NOT_FOUND;
 import static com.github.durex.shared.exceptions.model.ErrorCode.SAVE_ERROR;
 
-import com.github.durex.music.api.Music;
-import com.github.durex.music.api.PlayList;
-import com.github.durex.music.api.PlayListMusic;
+import com.github.durex.music.model.Music;
+import com.github.durex.music.model.PlayList;
+import com.github.durex.music.model.PlayListMusic;
 import com.github.durex.music.repository.PlayListMusicRepository;
 import com.github.durex.music.repository.PlayListRepository;
 import com.github.durex.shared.exceptions.ApiException;
