+++ core/music-reactive/src/test/java/com/github/durex/music/controller/MusicControllerTest.java
@@ -10,7 +10,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
-import com.github.durex.music.api.Music;
+import com.github.durex.music.model.Music;
 import com.github.durex.music.service.MusicService;
 import com.github.durex.music.support.DemoMusicData;
 import com.github.durex.utils.Json;
