+++ core/music-reactive/src/test/java/com/github/durex/music/service/MusicServiceTest.java
@@ -4,7 +4,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyString;
 
-import com.github.durex.music.api.Music;
+import com.github.durex.music.model.Music;
 import com.github.durex.music.repository.MusicRepository;
 import com.github.durex.music.support.DemoMusicData;
 import com.github.durex.shared.exceptions.ApiException;
