+++ core/music/src/main/java/com/github/durex/music/controller/MusicController.java
@@ -2,11 +2,11 @@ package com.github.durex.music.controller;
 
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 
-import com.github.durex.music.api.Music;
+import com.github.durex.music.model.Music;
 import com.github.durex.music.service.MusicService;
 import com.github.durex.shared.annotations.Logged;
 import com.github.durex.shared.annotations.Param;
-import com.github.durex.shared.api.RespData;
+import com.github.durex.shared.model.RespData;
 import com.github.durex.shared.utils.Helper;
 import java.util.List;
 import javax.enterprise.context.RequestScoped;
