+++ core/schema/music/entity/src/main/java/com/github/durex/music/support/DemoMusicData.java
@@ -1,8 +1,8 @@
 package com.github.durex.music.support;
 
-import com.github.durex.music.api.Music;
-import com.github.durex.music.api.PlayList;
-import com.github.durex.music.api.PlayListMusic;
+import com.github.durex.music.model.Music;
+import com.github.durex.music.model.PlayList;
+import com.github.durex.music.model.PlayListMusic;
 import com.github.durex.uuid.UniqID;
 import java.util.ArrayList;
 import java.util.List;
