+++ core/schema/music/r2dbc/src/main/java/com/github/durex/music/repository/CreatorPlayListRepository.java
@@ -1,10 +1,10 @@
 package com.github.durex.music.repository;
 
-import static com.github.durex.api.tables.QCreatorPlaylist.CREATOR_PLAYLIST;
-import static com.github.durex.api.tables.QPlaylist.PLAYLIST;
+import static com.github.durex.model.tables.QCreatorPlaylist.CREATOR_PLAYLIST;
+import static com.github.durex.model.tables.QPlaylist.PLAYLIST;
 
-import com.github.durex.api.tables.records.RPlaylist;
-import com.github.durex.music.api.PlayList;
+import com.github.durex.model.tables.records.RPlaylist;
+import com.github.durex.music.model.PlayList;
 import com.github.durex.music.mapper.PlayListMapper;
 import java.util.List;
 import java.util.stream.Collectors;
