+++ core/schema/music/r2dbc/src/main/java/com/github/durex/music/repository/MusicRepository.java
@@ -1,8 +1,8 @@
 package com.github.durex.music.repository;
 
-import static com.github.durex.api.tables.QMusic.MUSIC;
+import static com.github.durex.model.tables.QMusic.MUSIC;
 
-import com.github.durex.music.api.Music;
+import com.github.durex.music.model.Music;
 import com.github.durex.music.mapper.MusicMapper;
 import com.github.durex.sqlbuilder.SqlHelper;
 import com.github.durex.sqlbuilder.enums.WildCardType;
