+++ core/schema/music/r2dbc/src/test/java/com/github/durex/music/repository/PlayListRepositoryIT.java
@@ -2,7 +2,7 @@ package com.github.durex.music.repository;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.github.durex.music.api.PlayList;
+import com.github.durex.music.model.PlayList;
 import com.github.durex.music.support.DemoMusicData;
 import com.github.durex.sqlbuilder.enums.WildCardType;
 import com.github.durex.uuid.UniqID;
