+++ core/schema/music/repo/src/main/java/com/github/durex/music/repository/MusicRepository.java
@@ -1,17 +1,18 @@
 package com.github.durex.music.repository;
 
-import static com.github.durex.api.tables.QMusic.MUSIC;
+import static com.github.durex.model.tables.QMusic.MUSIC;
 
-import com.github.durex.music.api.Music;
-import com.github.durex.music.mapper.MusicMapper;
-import com.github.durex.sqlbuilder.SqlHelper;
-import com.github.durex.sqlbuilder.enums.WildCardType;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
+
+import com.github.durex.music.mapper.MusicMapper;
+import com.github.durex.music.model.Music;
+import com.github.durex.sqlbuilder.SqlHelper;
+import com.github.durex.sqlbuilder.enums.WildCardType;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.jooq.Condition;
