+++ core/schema/music/repo/src/test/java/com/github/durex/music/repository/CreatorPlayListRepositoryIT.java
@@ -4,8 +4,8 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.github.durex.music.api.Music;
-import com.github.durex.music.api.PlayList;
+import com.github.durex.music.model.Music;
+import com.github.durex.music.model.PlayList;
 import com.github.durex.music.support.DemoMusicData;
 import com.github.durex.uuid.UniqID;
 import io.quarkus.test.junit.QuarkusTest;
